fix unit storage in store
