+++ src/components/blocks/SetPointSimple/default.vue
@@ -1,3 +1,53 @@
+<script lang="ts">
+import Component from 'vue-class-component';
+
+import BlockComponent from '../BlockComponent';
+
+import { getById } from '@/store/blocks/SetPointSimple/getters';
+import { persist } from '@/store/blocks/SetPointSimple/actions';
+
+/* eslint-disable indent */
+@Component({
+  props: {
+    id: {
+      default: '',
+      type: String,
+    },
+  },
+})
+/* eslint-enable */
+export default class SetPointSimple extends BlockComponent {
+  valueInput = 0;
+
+  get blockData() {
+    return getById(this.$store, this.$props.id);
+  }
+
+  get setting() {
+    return this.blockData.setting;
+  }
+
+  get changed() {
+    return this.setting.value !== this.valueInput;
+  }
+
+  mounted() {
+    this.valueInput = this.setting.value;
+  }
+
+  save() {
+    // update the value of the unit
+    this.setting.value = this.valueInput;
+
+    persist(this.$store, {
+      id: this.blockData.id,
+      serviceId: this.blockData.serviceId,
+      setting: this.setting,
+    });
+  }
+}
+</script>
+
 <template>
   <q-card>
     <q-card-title>SetPointSimple ({{ id }})</q-card-title>
@@ -23,6 +73,7 @@
               stack-label="Value"
               placeholder="Value of SetPoint"
               type="number"
+              :suffix="setting.unitNotation"
             />
           </q-item-main>
         </q-item>
@@ -30,5 +81,3 @@
     </q-card-main>
   </q-card>
 </template>
-
-<script lang="ts" src="./SetPointSimple.ts"></script>
